:truck: api: move private certificate serializer from lib to src
