+++ api/src/certification/results/application/certification-controller.js
@@ -1,5 +1,5 @@
-import * as privateCertificateSerializer from '../../../../lib/infrastructure/serializers/jsonapi/private-certificate-serializer.js';
 import { usecases } from '../domain/usecases/index.js';
+import * as privateCertificateSerializer from '../infrastructure/serializers/private-certificate-serializer.js';
 
 const findUserCertifications = async function (request) {
   const userId = request.auth.credentials.userId;
