+++ api/src/certification/results/infrastructure/serializers/private-certificate-serializer.js
